[SPARK-28647][WEBUI][2.4] Recover additional metric feature
This PR is for backporting SPARK-28647(#25374) to branch-2.4.
The original PR removed `additional-metrics.js` but branch-2.4 still uses it so I don't remove it and related things for branch-2.4.

### What changes were proposed in this pull request?
Added checkboxes to enable users to select which optional metrics (`On Heap Memory`, `Off Heap Memory` and `Select All` in this case) to be shown in `ExecuorPage`.

### Why are the changes needed?
By SPARK-17019, `On Heap Memory` and `Off Heap Memory` are introduced as optional metrics. But they are not displayed because they are made `display: none` in css and there are no way to appear them.

### Does this PR introduce any user-facing change?
The previous `ExecutorPage` doesn't show optional metrics.
This change adds checkboxes to `ExecutorPage` for optional metrics.
We can choose which metrics should be shown by checking corresponding checkboxes.
![Screenshot from 2019-08-18 03-56-09](https://user-images.githubusercontent.com/4736016/63216148-2bfadb80-c16c-11e9-81e1-e1e66198dd6c.png)

### How was this patch tested?
Manual test.

Closes #25484 from sarutak/backport-SPARK-28647-branch-2.4.

Authored-by: Kousuke Saruta <sarutak@oss.nttdata.com>
Signed-off-by: Dongjoon Hyun <dhyun@apple.com>